Another progress in routing.
