+++ src/helpers/utils.js
@@ -12,3 +12,9 @@ export function isItDesktop() {
         return false;
     }
 }
+
+export function dayOfNum(number) {
+    let cases = [2, 0, 1, 1, 1, 2]; 
+    let titles = ['день', 'дня', 'дней'];
+    return titles[ (number%100>4 && number%100<20)? 2 : cases[(number%10<5)?number%10:5] ];
+}
